requirements.txt
django==2.0
psycopg2-binary
Pillow==5.3.0
djangorestframework==3.9.0
django-rest-swagger==2.2.0
django-autofixture==0.12.1
django-filter==2.1.0
drf-nested-routers==0.90.0
